refactoring :: ReviewsDomainInterface Transfer 나눔
