+++ Projects/Domain/ReviewsDomain/Interface/DTO/Response/QnaResponseDTO.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public struct QnaResponseDTO: Codable {
+    public let question, answer, area: String
+
+    public init(question: String, answer: String, area: String) {
+        self.question = question
+        self.answer = answer
+        self.area = area
+    }
+}
