+++ Projects/Domain/ReviewsDomain/Interface/DTO/Response/ReviewResponseDTO.swift
@@ -1,25 +1,25 @@
 import Foundation
 
-public struct ReviewDetailResponseDTO: Codable {
+public struct ReviewResponseDTO: Codable {
+    public let reviewID: String
     public let year: Int
     public let writer, createdDate: String
-    public let qnaResponses: [QnaResponseDTO]
 
     public init(
+        reviewID: String,
         year: Int,
         writer: String,
-        createdDate: String,
-        qnaResponses: [QnaResponseDTO]
+        createdDate: String
     ) {
+        self.reviewID = reviewID
         self.year = year
         self.writer = writer
         self.createdDate = createdDate
-        self.qnaResponses = qnaResponses
     }
 
     enum CodingKeys: String, CodingKey {
+        case reviewID = "review_id"
         case year, writer
         case createdDate = "created_date"
-        case qnaResponses = "qna_responses"
     }
 }
