+++ Projects/Domain/ReviewsDomain/Interface/Entity/ReviewDetailEntity.swift
@@ -18,12 +18,3 @@ public struct ReviewDetailEntity: Equatable {
     }
 }
 
-public struct QnaEntity: Equatable {
-    public let question, answer, area: String
-
-    public init(question: String, answer: String, area: String) {
-        self.question = question
-        self.answer = answer
-        self.area = area
-    }
-}
